refactor(ratewise): design-tokens 設定模組化拆分
- 1340 行單檔拆為 design-tokens/ 目錄（colors/layout/scale/components/tailwind + barrel）
- 各子檔皆 <600 行；import 路徑經 barrel 維持不變
- 同步 UpdatePrompt 守門測試改讀目錄全檔合併內容

測試：typecheck、lint、test 2513 passed、build:ratewise 全綠（tailwind config 解析 barrel）

via [Happy](https://happy.engineering)
