+++ .changeset/ratewise-design-tokens-modularize.md
@@ -1,0 +1,5 @@
+---
+'@app/ratewise': patch
+---
+
+內部重構：將 design tokens 設定模組化為多檔（colors/layout/scale/components/tailwind），提升可維護性，無使用者可見行為變更。
