+++ apps/ratewise/src/config/design-tokens/index.ts
@@ -1,0 +1,10 @@
+/**
+ * Design Token SSOT barrel。
+ * 實作依領域拆為 colors / layout / scale / components / tailwind 子檔，
+ * 對外維持單一 import 路徑 `config/design-tokens`。
+ */
+export * from './colors';
+export * from './layout';
+export * from './scale';
+export * from './components';
+export * from './tailwind';
